pyproject.toml
[project]
name = "prompt-architect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
